Kernel: Move Multi Processor Parser code to a separate directory
