+++ Kernel/Firmware/MultiProcessor/Parser.cpp
@@ -7,8 +7,8 @@
 
 #include <AK/StringView.h>
 #include <Kernel/Debug.h>
-#include <Kernel/Firmware/ACPI/MultiProcessorParser.h>
 #include <Kernel/Firmware/BIOS.h>
+#include <Kernel/Firmware/MultiProcessor/Parser.h>
 #include <Kernel/Interrupts/IOAPIC.h>
 #include <Kernel/Memory/TypedMapping.h>
 #include <Kernel/Sections.h>
